Harden the spool and stop three silent-corruption paths in aggregation
Independent review of the aggregation core reproduced two defects, and both are
fixed with tests that fail without the fix.

The DDSketch value estimate is converted from float64 without a bound, so a
device reporting LatencyNanos = MaxInt64 published a p95 of -9223372036854775808
into Redis, Postgres and the MCP tool. The estimate is clamped. Quantile also
accepted NaN, because both of its range comparisons are false for NaN, and
returned a silent zero; it now rejects it. Sketch.Sum had no caller anywhere and
could overflow, so it is gone with the field behind it.

Event.Bytes was never validated, so a negative byte count drained the per-key
counter below zero and MinInt64 wrapped it. It is rejected at the same boundary
that already rejects a negative latency.

Two liveness and observability gaps are closed. A window that can never be
published used to break the emit loop, so every window ordered behind it was
never published either; the loop now attempts all of them, commits the successes
and returns the first failure, which the existing partial-commit path already
supports. And shedding - the one path that drops telemetry to protect memory -
incremented a counter nothing ever read; it now reaches the metrics recorder and
the registry, and the pre-registration test caught the missing entry.

The spool's own findings, from a second review:

- recovery indexed any *.batch name, so a symlink planted in a shared spool
  directory was read and shipped to the edge. Only regular files are indexed,
  and Peek re-checks before reading.
- Enqueue wrote without fsync, so a crash could leave a durable name over torn
  bytes, which the edge then rejects as permanent and the drain discards. Writes
  and the directory entry are now synced.
- the directory was created 0700 but never corrected if it already existed, so a
  reused 0755 directory left telemetry world-readable.
- a failed unlink still removed the entry from the index, so a later restart
  resent a batch that had already been accepted and the byte budget overcounted
  free space. Removal now reports failure and accounting only follows the disk.
- recovery adopted whatever it found, ignoring maxBytes; it now trims.
- a zero-length payload consumed no budget but still consumed an inode and an
  index slot, so the index now has a ceiling derived from the budget.
- the sequence field was six digits, so past a million enqueues a restart sorted
  a newer batch ahead of an older one. Ordering now uses the parsed fields.

Each spool guarantee is pinned by a test verified to fail against the previous
code: symlink recovery, overfull recovery, the index ceiling, and the emission
stall.

+++ agent/capture_ebpf.go
@@ -30,8 +30,8 @@ func newEBPFSource(settings captureSettings) (domain.EventSource, error) {
 	}, ebpf.Settings{
 		MinLatency:               settings.minLatency,
 		MaxEventsPerCPUPerSecond: settings.maxEventsPerCPU,
-		SampleModulus:            uint32(settings.sampleModulus),
-		TargetTGID:               uint32(settings.targetTGID),
+		SampleModulus:            settings.sampleModulus,
+		TargetTGID:               settings.targetTGID,
 	})
 }
